+++ test/apollovm_languages_basic_test.dart
@@ -6,9 +6,9 @@ Future<void> main() async {
   print('BASIC TESTS DEFINITIONS');
 
   var definitions = <TestDefinition>[
-    TestDefinition('dart_basic_main_2.test.xml', '''
+    TestDefinition('dart_basic_main_1.test.xml', '''
 <?xml version="1.0" encoding="UTF-8"?>
-<test title="Basic main(List<String>) 2">
+<test title="Basic main(List<String>) 1">
     <source language="dart">
         <![CDATA[
             void main(List<String> args) {
@@ -52,9 +52,9 @@ Future<void> main() async {
 ]]></source-generated>
 </test>
     '''),
-    TestDefinition('java11_basic_main_2.test.xml', '''
+    TestDefinition('java11_basic_main_1.test.xml', '''
 <?xml version="1.0" encoding="UTF-8"?>
-<test title="Basic main(Object[]) 2">
+<test title="Basic main(Object[]) 1">
     <source language="java11">
         <![CDATA[
 
@@ -127,6 +127,121 @@ class Foo {
 ]]></source-generated>
 </test>
   '''),
+    TestDefinition('dart_basic_main_2.test.xml', '''
+<?xml version="1.0" encoding="UTF-8"?>
+<test title="Basic main(List<String>) 2">
+    <source language="dart">
+        <![CDATA[
+            void main(List args) {
+              var title = args[0];
+              var a = args[1];
+              var b = args[2];
+              var c = args[3];
+              var sumAB = a + b ;
+              var sumABC = a + b + c;
+              print(title);
+              print(sumAB);
+              print(sumABC);
+              
+              // List:
+              var list = <int>[a, b, c];
+              print(list);
+              
+              var listEmpty = <String>[];
+              print(listEmpty);
+            }
+        ]]>
+    </source>
+    <call function="main">
+        [
+          ["Integers:", 10, 20, 30]
+        ]
+    </call>
+    <output>
+         ["Integers:", 30, 60, [10, 20, 30], []]
+    </output>
+    <source-generated language="dart"><![CDATA[<<<< [SOURCES_BEGIN] >>>>
+<<<< NAMESPACE="" >>>>
+<<<< CODE_UNIT_START="/test" >>>>
+  void main(List<dynamic> args) {
+    var title = args[0];
+    var a = args[1];
+    var b = args[2];
+    var c = args[3];
+    var sumAB = a + b;
+    var sumABC = a + b + c;
+    print(title);
+    print(sumAB);
+    print(sumABC);
+    var list = <int>[a, b, c];
+    print(list);
+    var listEmpty = <String>[];
+    print(listEmpty);
+  }
+
+<<<< CODE_UNIT_END="/test" >>>>
+<<<< [SOURCES_END] >>>>
+]]></source-generated>
+</test>
+    '''),
+    TestDefinition('dart_basic_main_3.test.xml', '''
+<?xml version="1.0" encoding="UTF-8"?>
+<test title="Basic main(List<String>) 3">
+    <source language="dart">
+        <![CDATA[
+            void main(List args) {
+              var title = args[0];
+              var a = args[1];
+              var b = args[2];
+              var c = args[3];
+              var sumAB = a + b ;
+              var sumABC = a + b + c;
+              print(title);
+              print(sumAB);
+              print(sumABC);
+              
+              // Map:
+              var map = <String,int>{
+              'a': a,
+              'b': b,
+              'c': c,
+              };
+              print(map);
+              print(map['b']);
+            }
+        ]]>
+    </source>
+    <call function="main">
+        [
+          ["Integers:", 10, 20, 30]
+        ]
+    </call>
+    <output>
+         ["Integers:", 30, 60, {"a": 10, "b": 20, "c": 30}, 20]
+    </output>
+    <source-generated language="dart"><![CDATA[<<<< [SOURCES_BEGIN] >>>>
+<<<< NAMESPACE="" >>>>
+<<<< CODE_UNIT_START="/test" >>>>
+  void main(List<dynamic> args) {
+    var title = args[0];
+    var a = args[1];
+    var b = args[2];
+    var c = args[3];
+    var sumAB = a + b;
+    var sumABC = a + b + c;
+    print(title);
+    print(sumAB);
+    print(sumABC);
+    var map = <String,int>{'a': a, 'b': b, 'c': c};
+    print(map);
+    print(map['b']);
+  }
+
+<<<< CODE_UNIT_END="/test" >>>>
+<<<< [SOURCES_END] >>>>
+]]></source-generated>
+</test>
+    '''),
   ];
 
   await runTestDefinitions(definitions);